Updated to allow pulling from PROD database,
Also moved to permanent AWS account.

Right now, the dev is just commented out and you have to alternate by
commenting out the version you don't want. (If you want to run dev, you
got to comment out prod). I am thinking it might be more efficient from
a developer standpoint to use an if statement, but from an actual
running standpoint, I don't know if this is the right way in production
as it technically slows down the code, however slight.